requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
asgiref=3.8.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.1.3=pypi_0
beautifulsoup4=4.12.3=pypi_0
bs4=0.0.2=pypi_0
build=1.2.1=pypi_0
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.2.2=h56e8100_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.5.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_0
dataclasses-json=0.6.6=pypi_0
debugpy=1.6.7=py310hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
deepdiff=7.0.1=pypi_0
deprecated=1.2.14=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.1.1=pypi_0
emoji=2.12.1=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
filelock=3.14.0=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.30.0=pypi_0
googleapis-common-protos=1.63.1=pypi_0
gpt4all=2.7.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.64.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.3=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=7.1.0=hd8ed1ab_0
install=1.3.5=pypi_0
ipykernel=6.29.3=pyha63f2e9_0
ipython=8.22.2=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpath-python=1.0.6=pypi_0
jsonpointer=2.4=pypi_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py310h5588dad_0
kubernetes=30.1.0=pypi_0
langchain=0.2.3=pypi_0
langchain-chroma=0.1.1=pypi_0
langchain-community=0.2.4=pypi_0
langchain-core=0.2.5=pypi_0
langchain-openai=0.1.8=pypi_0
langchain-text-splitters=0.2.1=pypi_0
langchainhub=0.1.18=pypi_0
langdetect=1.0.9=pypi_0
langsmith=0.1.75=pypi_0
libffi=3.4.4=hd77b12b_0
libsodium=1.0.18=h8d14728_1
lxml=5.2.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nltk=3.8.1=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.18.0=pypi_0
openai=1.32.1=pypi_0
openssl=3.0.13=h2bbff1b_0
opentelemetry-api=1.25.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.25.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.25.0=pypi_0
opentelemetry-instrumentation=0.46b0=pypi_0
opentelemetry-instrumentation-asgi=0.46b0=pypi_0
opentelemetry-instrumentation-fastapi=0.46b0=pypi_0
opentelemetry-proto=1.25.0=pypi_0
opentelemetry-sdk=1.25.0=pypi_0
opentelemetry-semantic-conventions=0.46b0=pypi_0
opentelemetry-util-http=0.46b0=pypi_0
ordered-set=4.1.0=pypi_0
orjson=3.10.1=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pinecone-client=4.1.1=pypi_0
pinecone-plugin-interface=0.0.7=pypi_0
pip=23.3.1=py310haa95532_0
platformdirs=4.2.1=pyhd8ed1ab_0
posthog=3.5.0=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pypdf=4.2.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
pyreadline3=3.4.1=pypi_0
python=3.10.14=he1021f5_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pypi_0
python-iso639=2024.4.27=pypi_0
python-magic=0.4.27=pypi_0
python-multipart=0.0.9=pypi_0
python_abi=3.10=2_cp310
pywin32=305=py310h2bbff1b_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py310hd77b12b_0
rapidfuzz=3.9.3=pypi_0
regex=2024.4.28=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.4.2=pypi_0
scipy=1.13.0=pypi_0
setuptools=68.2.2=py310haa95532_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.30=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
sympy=1.12.1=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
tornado=6.2=py310he2412df_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typer=0.12.3=pypi_0
types-requests=2.32.0.20240602=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024a=h04d1e81_0
ujson=5.10.0=pypi_0
unstructured=0.14.4=pypi_0
unstructured-client=0.23.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.30.1=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchfiles=0.22.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websocket-client=1.8.0=pypi_0
websockets=12.0=pypi_0
wheel=0.41.2=py310haa95532_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h8cc25b3_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=hd77b12b_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
